requirements.txt
numpy~=1.22.4
pandas~=1.4.3
matplotlib~=3.5.2
torch~=1.13.1
tqdm~=4.64.0
librosa~=0.9.2
fast-bss-eval~=0.1.4
PyYAML~=6.0
torchaudio~=0.13.1
pytorch-lightning~=1.9.5
HyperPyYAML~=1.0.1
torchmetrics~=0.10.3
asteroid~=0.6.0
